feat: rename clas add keyword use case  and formating file
